Fix tset builds by specifying targets properly.
ic_count.c should not be qualified; it is an output file and so it goes
into ${.OBJDIR} implicitly.

TS_DATA files originate in the source tree, not the object tree, so don't
look for it there.  Unconditionally generate the TS_DATA targets and let
make decide whether they need to run.  Instead of copying them to .OBJDIR,
just gzcat them directly to their target in .OBJDIR.